refactor: move loop_if, element_if, composite_if to sibling modules
Step 4 of the map_if refactor plan. Three of the five inline classes
in map_if/__init__.py move to siblings:

- pyx12/map_if/_element.py     (element_if — leaf validator)
- pyx12/map_if/_composite.py   (composite_if — sub-element container)
- pyx12/map_if/_loop.py        (loop_if — recursive loop/segment container)

__init__.py shrinks from 1633 -> 855 lines, retains only `map_if`,
`segment_if`, and `load_map_file` for now (PR 5 will move those).
The package re-exports the moved classes so `pyx12.map_if.loop_if`
etc. remain importable — no public API change.

### Circular-import handling
- `_element.py` is a pure leaf — only imports from `_base`.
- `_composite.py` instantiates `element_if` -> imports `_element` at
  module top. No cycle (element doesn't reference composite).
- `_loop.py` instantiates `segment_if` (still in `__init__.py`) and
  recursively `loop_if` (self). Self-reference is class-scope, no
  import needed. For `segment_if`, a local `from . import segment_if`
  inside `loop_if.__init__` breaks the cycle: by the time `__init__`
  runs (under `load_map_file`), the package is fully loaded so the
  attribute lookup succeeds. mypy resolves the name via a
  TYPE_CHECKING import at module top.

After PR 5 (which moves `segment_if` into its own module), the
runtime import inside `loop_if.__init__` becomes a regular top-level
`from ._segment import segment_if`.

### pyx12.pyproj updated
Three new `<Compile Include>` entries for the new files.

### Verification
- pytest: 536 passed
- mypy --strict: clean (79 source files, was 76)
- ruff format + ruff check --select I: clean
- Import smoke: all 9 names still resolvable from `pyx12.map_if`
