+++ pyx12/map_if/__init__.py
@@ -223,292 +223,7 @@ class map_if(x12_node):
                         yield c
 
 
-############################################################
-# Loop Interface
-############################################################
-class loop_if(x12_node):
-    """
-    Loop Interface
-    """
-
-    root: Any
-    pos_map: dict[int, list[Any]]
-    base_name: str
-    _cur_count: int
-    type: str | None
-    usage: str | None
-    pos: int
-    repeat: str | None
-
-    def __init__(self, root: Any, parent: Any, elem: Element) -> None:
-        """ """
-        x12_node.__init__(self)
-        self.root = root
-        self.parent = parent
-        self.pos_map = {}
-        self.base_name = "loop"
-        self._cur_count = 0
-
-        self.id = elem.get("xid")
-        self.path = self.id or ""
-        self.type = elem.get("type")
-
-        self.name = elem.get("name") if elem.get("name") else elem.findtext("name")
-        self.usage = elem.get("usage") if elem.get("usage") else elem.findtext("usage")
-        self.pos = int(_required_attr(elem, "pos"))
-        self.repeat = elem.get("repeat") if elem.get("repeat") else elem.findtext("repeat")
-
-        for e in elem.findall("loop"):
-            loop_node = loop_if(self.root, self, e)
-            try:
-                self.pos_map[loop_node.pos].append(loop_node)
-            except KeyError:
-                self.pos_map[loop_node.pos] = [loop_node]
-        for e in elem.findall("segment"):
-            seg_node = segment_if(self.root, self, e)
-            try:
-                self.pos_map[seg_node.pos].append(seg_node)
-            except KeyError:
-                self.pos_map[seg_node.pos] = [seg_node]
-
-        # For the segments with duplicate ordinals, adjust the path to be unique
-        for ord1 in sorted(self.pos_map):
-            if len(self.pos_map[ord1]) > 1:
-                for seg_node in [n for n in self.pos_map[ord1] if n.is_segment()]:
-                    id_elem = seg_node.guess_unique_key_id_element()
-                    if id_elem is not None:
-                        seg_node.path = seg_node.path + "[" + id_elem.valid_codes[0] + "]"
-
-    def debug_print(self) -> None:
-        sys.stdout.write(self.__repr__())
-        for ord1 in sorted(self.pos_map):
-            for node in self.pos_map[ord1]:
-                node.debug_print()
-
-    def __len__(self) -> int:
-        i = 0
-        for ord1 in sorted(self.pos_map):
-            i += len(self.pos_map[ord1])
-        return i
-
-    def __repr__(self) -> str:
-        """
-        :rtype: string
-        """
-        out = ""
-        if self.id:
-            out += "LOOP %s" % (self.id)
-        if self.name:
-            out += '  "%s"' % (self.name)
-        if self.usage:
-            out += "  usage: %s" % (self.usage)
-        if self.pos:
-            out += "  pos: %s" % (self.pos)
-        if self.repeat:
-            out += "  repeat: %s" % (self.repeat)
-        out += "\n"
-        return out
-
-    def get_max_repeat(self) -> int:
-        if self.repeat is None:
-            return MAXINT
-        if self.repeat == "&gt;1" or self.repeat == ">1":
-            return MAXINT
-        return int(self.repeat)
-
-    def get_parent(self) -> Any:
-        return self.parent
-
-    def get_first_node(self) -> Any:
-        pos_keys = sorted(self.pos_map)
-        if len(pos_keys) > 0:
-            return self.pos_map[pos_keys[0]][0]
-        else:
-            return None
-
-    def get_first_seg(self) -> Any:
-        first = self.get_first_node()
-        if first.is_segment():
-            return first
-        else:
-            return None
-
-    def childIterator(self) -> Iterator[Any]:
-        for ord1 in sorted(self.pos_map):
-            for child in self.pos_map[ord1]:
-                yield child
-
-    def getnodebypath(self, spath: str) -> Any:
-        """
-        :param spath: remaining path to match
-        :type spath: string
-        :return: matching node, or None is no match
-        """
-        pathl = spath.split("/")
-        if len(pathl) == 0:
-            return None
-        for ord1 in sorted(self.pos_map):
-            for child in self.pos_map[ord1]:
-                if child.is_loop():
-                    if child.id.upper() == pathl[0].upper():
-                        if len(pathl) == 1:
-                            return child
-                        else:
-                            return child.getnodebypath("/".join(pathl[1:]))
-                elif child.is_segment() and len(pathl) == 1:
-                    if pathl[0].find("[") == -1:  # No id to match
-                        if pathl[0] == child.id:
-                            return child
-                    else:
-                        seg_id = pathl[0][0 : pathl[0].find("[")]
-                        id_val = pathl[0][pathl[0].find("[") + 1 : pathl[0].find("]")]
-                        if seg_id == child.id:
-                            possible = child.get_unique_key_id_element(id_val)
-                            if possible is not None:
-                                return child
-        raise EngineError('getnodebypath failed. Path "%s" not found' % spath)
-
-    def getnodebypath2(self, path_str: str) -> Any:
-        """
-        Try x12 path
-
-        :param path_str: remaining path to match
-        :type path_str: string
-        :return: matching node, or None is no match
-        """
-        x12path = X12Path(path_str)
-        if x12path.empty():
-            return None
-        for ord1 in sorted(self.pos_map):
-            for child in self.pos_map[ord1]:
-                if child.is_loop() and len(x12path.loop_list) > 0:
-                    if child.id.upper() == x12path.loop_list[0].upper():
-                        if len(x12path.loop_list) == 1 and x12path.seg_id is None:
-                            return child
-                        else:
-                            del x12path.loop_list[0]
-                            return child.getnodebypath2(x12path.format())
-                elif (
-                    child.is_segment()
-                    and len(x12path.loop_list) == 0
-                    and x12path.seg_id is not None
-                ):
-                    if x12path.id_val is None:
-                        if x12path.seg_id == child.id:
-                            return child.getnodebypath2(x12path.format())
-                    else:
-                        seg_id = x12path.seg_id
-                        id_val = x12path.id_val
-                        if seg_id == child.id:
-                            possible = child.get_unique_key_id_element(id_val)
-                            if possible is not None:
-                                return child.getnodebypath2(x12path.format())
-        raise EngineError('getnodebypath2 failed. Path "%s" not found' % path_str)
-
-    def get_child_count(self) -> int:
-        return self.__len__()
-
-    def get_child_node_by_idx(self, idx: int) -> Any:
-        """
-        :param idx: zero based
-        """
-        raise EngineError("loop_if.get_child_node_by_idx is not a valid call for a loop_if")
-
-    def get_seg_count(self) -> int:
-        """
-        :return: Number of child segments
-        :rtype: integer
-        """
-        i = 0
-        for ord1 in sorted(self.pos_map):
-            for child in self.pos_map[ord1]:
-                if child.is_segment():
-                    i += 1
-        return i
-
-    def is_loop(self) -> bool:
-        """
-        :rtype: boolean
-        """
-        return True
-
-    def is_match(self, seg_data: pyx12.segment.Segment) -> bool:
-        """
-        :type seg_data: L{segment<segment.Segment>}
-        :return: Is the segment a match to this loop?
-        :rtype: boolean
-        """
-        pos_keys = sorted(self.pos_map)
-        child = self.pos_map[pos_keys[0]][0]
-        if child.is_loop():
-            return bool(child.is_match(seg_data))
-        elif child.is_segment():
-            if child.is_match(seg_data):
-                return True
-            else:
-                return False  # seg does not match the first segment in loop, so not valid
-        else:
-            return False
-
-    def get_child_seg_node(self, seg_data: pyx12.segment.Segment) -> Any:
-        """
-        Return the child segment matching the segment data
-        """
-        for child in self.childIterator():
-            if child.is_segment() and child.is_match(seg_data):
-                return child
-        return None
-
-    def get_child_loop_node(self, seg_data: pyx12.segment.Segment) -> Any:
-        """
-        Return the child segment matching the segment data
-        """
-        for child in self.childIterator():
-            if child.is_loop() and child.is_match(seg_data):
-                return child
-        return None
-
-    def get_cur_count(self) -> int:
-        """
-        :return: current count
-        :rtype: int
-        """
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def incr_cur_count(self) -> None:
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def reset_child_count(self) -> None:
-        """
-        Set cur_count of child nodes to zero
-        """
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def reset_cur_count(self) -> None:
-        """
-        Set cur_count of node and child nodes to zero
-        """
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def set_cur_count(self, ct: int) -> None:
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def get_counts_list(self, ct_list: list[tuple[str, int]]) -> bool:
-        """
-        Build a list of (path, ct) of the current node and parents
-        Gets the node counts to apply to another map
-        :param ct_list: List to append to
-        :type ct_list: list[(string, int)]
-        """
-        raise DeprecationWarning("Moved to nodeCounter")
-
-    def loop_segment_iterator(self) -> Iterator[Any]:
-        yield self
-        for ord1 in sorted(self.pos_map):
-            for child in self.pos_map[ord1]:
-                if child.is_loop() or child.is_segment():
-                    for c in child.loop_segment_iterator():
-                        yield c
+from ._loop import loop_if  # noqa: E402, I001
 
 
 class segment_if(x12_node):
@@ -945,498 +660,8 @@ class segment_if(x12_node):
         yield self
 
 
-############################################################
-# Element Interface
-############################################################
-class element_if(x12_node):
-    """
-    Element Interface
-    """
-
-    root: Any
-    base_name: str
-    valid_codes: list[str | None]
-    external_codes: str | None
-    rec: re.Pattern[str] | None
-    refdes: str | None
-    data_ele: str | None
-    _data_ele: _DataEle | None
-    usage: str | None
-    seq: int
-    max_use: str | None
-    res: str | None
-
-    def __init__(self, root: Any, parent: Any, elem: Element) -> None:
-        """
-        :param parent: parent node
-        """
-        x12_node.__init__(self)
-        self.children = []
-        self.root = root
-        self.parent = parent
-        self.base_name = "element"
-        self.valid_codes = []
-        self.external_codes = None
-        self.rec = None
-
-        self.id = elem.get("xid")
-        self.refdes = self.id
-        self.data_ele = elem.get("data_ele") if elem.get("data_ele") else elem.findtext("data_ele")
-        # Eagerly cache the data element definition; a map that references an
-        # undefined data_ele fails at validation time, matching legacy behavior.
-        try:
-            self._data_ele = (
-                self.root.data_elements.get_by_elem_num(self.data_ele) if self.data_ele else None
-            )
-        except EngineError:
-            self._data_ele = None
-        self.usage = elem.get("usage") if elem.get("usage") else elem.findtext("usage")
-        self.name = elem.get("name") if elem.get("name") else elem.findtext("name")
-        self.seq = int(_required_attr(elem, "seq"))
-        self.path = elem.get("seq") if elem.get("seq") else (elem.findtext("seq") or "")  # type: ignore[assignment]
-        self.max_use = elem.get("max_use") if elem.get("max_use") else elem.findtext("max_use")
-        self.res = elem.findtext("regex")
-        try:
-            if self.res is not None and self.res != "":
-                self.rec = re.compile(self.res, re.S)
-        except Exception:
-            raise EngineError('Element regex "%s" failed to compile' % (self.res))
-
-        v = elem.find("valid_codes")
-        if v is not None:
-            self.external_codes = v.get("external")
-            for c in v.findall("code"):
-                self.valid_codes.append(c.text)
-
-    def debug_print(self) -> None:
-        sys.stdout.write(self.__repr__())
-        for node in self.children:
-            node.debug_print()
-
-    def __repr__(self) -> str:
-        """
-        :rtype: string
-        """
-        data_ele = self._resolve_data_ele()
-        out = '%s "%s"' % (self.refdes, self.name)
-        if self.data_ele:
-            out += "  data_ele: %s" % (self.data_ele)
-        if self.usage:
-            out += "  usage: %s" % (self.usage)
-        if self.seq:
-            out += "  seq: %i" % (self.seq)
-        out += "  %s(%i, %i)" % (data_ele["data_type"], data_ele["min_len"], data_ele["max_len"])
-        if self.external_codes:
-            out += "   external codes: %s" % (self.external_codes)
-        out += "\n"
-        return out
-
-    def _resolve_data_ele(self) -> _DataEle:
-        if self._data_ele is not None:
-            return self._data_ele
-        return cast(_DataEle, self.root.data_elements.get_by_elem_num(self.data_ele))
-
-    def _error(self, errh: Any, err_str: str, err_cde: str, elem_val: str | None) -> None:
-        """
-        Forward the error to an error_handler
-        """
-        errh.ele_error(err_cde, err_str, elem_val, self.refdes)
-
-    def _valid_code(self, code: str | None) -> bool:
-        """
-        Verify the x12 element value is in the given list of valid codes
-        :return: True if found, else False
-        :rtype: boolean
-        """
-        if code in self.valid_codes:
-            return True
-        return False
-
-    def get_parent(self) -> Any:
-        """
-        :return: ref to parent class instance
-        """
-        return self.parent
-
-    def is_match(self) -> bool:
-        """
-        :return:
-        :rtype: boolean
-        """
-        # match also by ID
-        raise NotImplementedError("Override in sub-class")
-
-    def is_valid(self, elem: Any, errh: Any, type_list: list[str | None] | None = None) -> bool:
-        """
-        Is this a valid element?
-
-        :param elem: element instance
-        :type elem: L{element<segment.Element>}
-        :param errh: instance of error_handler
-        :param type_list: Optional data/time type list
-        :type type_list: list[string]
-        :return: True if valid
-        :rtype: boolean
-        """
-        if type_list is None:
-            type_list = []
-        errh.add_ele(self)
-
-        if elem and elem.is_composite():
-            err_str = 'Data element "%s" (%s) is an invalid composite' % (self.name, self.refdes)
-            self._error(errh, err_str, "6", elem.__repr__())
-            return False
-        if elem is None or elem.get_value() == "":
-            return self._validate_when_empty(errh)
-        if self.usage == "N" and elem.get_value() != "":
-            err_str = 'Data element "%s" (%s) is marked as Not Used' % (self.name, self.refdes)
-            self._error(errh, err_str, "10", None)
-            return False
-
-        elem_val = elem.get_value()
-        valid = self._validate_length(elem_val, errh)
-        if not self._validate_control_chars(elem_val, errh):
-            return False  # control char errors trump later checks
-        valid &= self._validate_trailing_spaces(elem_val, errh)
-        valid &= self._is_valid_code(elem_val, errh)
-        valid &= self._validate_data_type(elem_val, errh)
-        if type_list:
-            valid &= self._validate_type_list(elem_val, type_list, errh)
-        valid &= self._validate_regex(elem_val, errh)
-        return bool(valid)
-
-    def _validate_when_empty(self, errh: Any) -> bool:
-        if self.usage in ("N", "S"):
-            return True
-        if self.usage == "R" and (
-            self.seq != 1 or not self.parent.is_composite() or self.parent.usage == "R"
-        ):
-            err_str = 'Mandatory data element "%s" (%s) is missing' % (self.name, self.refdes)
-            self._error(errh, err_str, "1", None)
-            return False
-        return True
-
-    def _validate_length(self, elem_val: str, errh: Any) -> bool:
-        data_ele = self._resolve_data_ele()
-        data_type = data_ele["data_type"]
-        min_len = data_ele["min_len"]
-        max_len = data_ele["max_len"]
-        # Numeric types ignore "-" and "." for length purposes.
-        if data_type is not None and (data_type == "R" or data_type[0] == "N"):
-            measured = elem_val.replace("-", "").replace(".", "")
-        else:
-            measured = elem_val
-        elem_len = len(measured)
-        valid = True
-        if elem_len < min_len:
-            err_str = 'Data element "%s" (%s) is too short: len("%s") = %i < %i (min_len)' % (
-                self.name,
-                self.refdes,
-                elem_val,
-                elem_len,
-                min_len,
-            )
-            self._error(errh, err_str, "4", elem_val)
-            valid = False
-        if elem_len > max_len:
-            err_str = 'Data element "%s" (%s) is too long: len("%s") = %i > %i (max_len)' % (
-                self.name,
-                self.refdes,
-                elem_val,
-                elem_len,
-                max_len,
-            )
-            self._error(errh, err_str, "5", elem_val)
-            valid = False
-        return valid
-
-    def _validate_control_chars(self, elem_val: str, errh: Any) -> bool:
-        res, bad_string = validation.contains_control_character(elem_val)
-        if not res:
-            return True
-        err_str = 'Data element "%s" (%s), contains an invalid control character(%s)' % (
-            self.name,
-            self.refdes,
-            bad_string,
-        )
-        self._error(errh, err_str, "6", bad_string)
-        return False
-
-    def _validate_trailing_spaces(self, elem_val: str, errh: Any) -> bool:
-        data_ele = self._resolve_data_ele()
-        if data_ele["data_type"] not in ("AN", "ID") or elem_val[-1] != " ":
-            return True
-        if len(elem_val.rstrip()) < data_ele["min_len"]:
-            return True
-        err_str = 'Data element "%s" (%s) has unnecessary trailing spaces. (%s)' % (
-            self.name,
-            self.refdes,
-            elem_val,
-        )
-        self._error(errh, err_str, "6", elem_val)
-        return False
-
-    def _validate_data_type(self, elem_val: str, errh: Any) -> bool:
-        data_type = self._resolve_data_ele()["data_type"]
-        if validation.IsValidDataType(
-            elem_val, cast(str, data_type), self.root.param.get("charset"), self.root.icvn
-        ):
-            return True
-        if data_type in ("RD8", "DT", "D8", "D6"):
-            err_str = 'Data element "%s" (%s) contains an invalid date (%s)' % (
-                self.name,
-                self.refdes,
-                elem_val,
-            )
-            self._error(errh, err_str, "8", elem_val)
-        elif data_type == "TM":
-            err_str = 'Data element "%s" (%s) contains an invalid time (%s)' % (
-                self.name,
-                self.refdes,
-                elem_val,
-            )
-            self._error(errh, err_str, "9", elem_val)
-        else:
-            err_str = 'Data element "%s" (%s) is type %s, contains an invalid character(%s)' % (
-                self.name,
-                self.refdes,
-                data_type,
-                elem_val,
-            )
-            self._error(errh, err_str, "6", elem_val)
-        return False
-
-    def _validate_type_list(self, elem_val: str, type_list: list[str | None], errh: Any) -> bool:
-        valid_type = False
-        for dtype in type_list:
-            if dtype is not None:
-                valid_type |= validation.IsValidDataType(
-                    elem_val, dtype, self.root.param.get("charset")
-                )
-        if valid_type:
-            return True
-        if "TM" in type_list:
-            err_str = 'Data element "%s" (%s) contains an invalid time (%s)' % (
-                self.name,
-                self.refdes,
-                elem_val,
-            )
-            self._error(errh, err_str, "9", elem_val)
-        elif any(t in type_list for t in ("RD8", "DT", "D8", "D6")):
-            err_str = 'Data element "%s" (%s) contains an invalid date (%s)' % (
-                self.name,
-                self.refdes,
-                elem_val,
-            )
-            self._error(errh, err_str, "8", elem_val)
-        return False
-
-    def _validate_regex(self, elem_val: str, errh: Any) -> bool:
-        if self.rec is None or self.rec.search(elem_val):
-            return True
-        err_str = 'Data element "%s" with a value of (%s)' % (self.name, elem_val)
-        err_str += ' failed to match the regular expression "%s"' % (self.res)
-        self._error(errh, err_str, "7", elem_val)
-        return False
-
-    def _is_valid_code(self, elem_val: str, errh: Any) -> bool:
-        """
-        :rtype: boolean
-        """
-        bValidCode = False
-        if len(self.valid_codes) == 0 and self.external_codes is None:
-            bValidCode = True
-        if elem_val in self.valid_codes:
-            bValidCode = True
-        if self.external_codes is not None and self.root.ext_codes.isValid(
-            self.external_codes, elem_val
-        ):
-            bValidCode = True
-        if not bValidCode:
-            err_str = "(%s) is not a valid code for %s (%s)" % (elem_val, self.name, self.refdes)
-            self._error(errh, err_str, "7", elem_val)
-            return False
-        return True
-
-    def get_data_type(self) -> str | None:
-        return self._resolve_data_ele()["data_type"]
-
-    @property
-    def data_type(self) -> str | None:
-        return self._resolve_data_ele()["data_type"]
-
-    @property
-    def min_len(self) -> int:
-        return self._resolve_data_ele()["min_len"]
-
-    @property
-    def max_len(self) -> int:
-        return self._resolve_data_ele()["max_len"]
-
-    @property
-    def data_element_name(self) -> str | None:
-        return self._resolve_data_ele()["name"]
-
-    def get_seg_count(self) -> None:
-        """ """
-        pass
-
-    def is_element(self) -> bool:
-        """
-        :rtype: boolean
-        """
-        return True
-
-    def get_path(self) -> str:
-        """
-        :return: path - XPath style
-        :rtype: string
-        """
-        if self._fullpath:
-            return self._fullpath
-        # get enclosing loop
-        parent_path = self.get_parent_segment().parent.get_path()
-        # add the segment, element, and sub-element path
-        self._fullpath = parent_path + "/" + (self.id or "")
-        return self._fullpath
-
-    def get_parent_segment(self) -> Any:
-        # pop to enclosing loop
-        p = self.parent
-        while not p.is_segment():
-            p = p.parent
-        return p
-
-
-############################################################
-# Composite Interface
-############################################################
-class composite_if(x12_node):
-    """
-    Composite Node Interface
-    """
-
-    root: Any
-    base_name: str
-    refdes: str | None
-    data_ele: str | None
-    usage: str | None
-    seq: int
-    repeat: int
-
-    def __init__(self, root: Any, parent: Any, elem: Element) -> None:
-        """
-        Get the values for this composite
-        :param parent: parent node
-        """
-        x12_node.__init__(self)
-
-        self.children = []
-        self.root = root
-        self.parent = parent
-        self.path = ""
-        self.base_name = "composite"
-
-        self.id = elem.get("xid")
-        self.refdes = elem.findtext("refdes") if elem.findtext("refdes") else self.id
-        self.data_ele = elem.get("data_ele") if elem.get("data_ele") else elem.findtext("data_ele")
-        self.usage = elem.get("usage") if elem.get("usage") else elem.findtext("usage")
-        self.seq = int(_required_attr(elem, "seq"))
-        if (r := elem.get("repeat")) is not None:
-            self.repeat = int(r)
-        elif (r := elem.findtext("repeat")) is not None:
-            self.repeat = int(r)
-        else:
-            self.repeat = 1
-        self.name = elem.get("name") if elem.get("name") else elem.findtext("name")
-
-        for e in elem.findall("element"):
-            self.children.append(element_if(self.root, self, e))
-
-    def _error(self, errh: Any, err_str: str, err_cde: str, elem_val: str) -> None:
-        """
-        Forward the error to an error_handler
-        """
-        err_str2 = err_str.replace("\n", "").replace("\r", "")
-        elem_val2 = elem_val.replace("\n", "").replace("\r", "")
-        errh.ele_error(err_cde, err_str2, elem_val2, self.refdes)
-
-    def debug_print(self) -> None:
-        sys.stdout.write(self.__repr__())
-        for node in self.children:
-            node.debug_print()
-
-    def __repr__(self) -> str:
-        """
-        :rtype: string
-        """
-        out = '%s "%s"' % (self.id, self.name)
-        if self.usage:
-            out += "  usage %s" % (self.usage)
-        if self.seq:
-            out += "  seq %i" % (self.seq)
-        if self.refdes:
-            out += "  refdes %s" % (self.refdes)
-        out += "\n"
-        return out
-
-    def xml(self) -> None:
-        """
-        Sends an xml representation of the composite to stdout
-        """
-        sys.stdout.write("<composite>\n")
-        for sub_elem in self.children:
-            sub_elem.xml()
-        sys.stdout.write("</composite>\n")
-
-    def is_valid(self, comp_data: Any, errh: Any) -> bool:
-        """
-        Validates the composite
-        :param comp_data: data composite instance, has multiple values
-        :param errh: instance of error_handler
-        :rtype: boolean
-        """
-        valid = True
-        if (comp_data is None or comp_data.is_empty()) and self.usage in ("N", "S"):
-            return True
-
-        if self.usage == "R":
-            good_flag = False
-            if comp_data is not None:
-                for sub_ele in comp_data:
-                    if sub_ele is not None and len(sub_ele.get_value()) > 0:
-                        good_flag = True
-                        break
-            if not good_flag:
-                err_str = 'At least one component of composite "%s" (%s) is required' % (
-                    self.name,
-                    self.refdes,
-                )
-                errh.ele_error("2", err_str, None, self.refdes)
-                return False
-
-        if self.usage == "N" and not comp_data.is_empty():
-            err_str = 'Composite "%s" (%s) is marked as Not Used' % (self.name, self.refdes)
-            errh.ele_error("5", err_str, None, self.refdes)
-            return False
-
-        if len(comp_data) > self.get_child_count():
-            err_str = 'Too many sub-elements in composite "%s" (%s)' % (self.name, self.refdes)
-            errh.ele_error("3", err_str, None, self.refdes)
-            valid = False
-        for i in range(min(len(comp_data), self.get_child_count())):
-            valid &= self.get_child_node_by_idx(i).is_valid(comp_data[i], errh)
-        for i in range(min(len(comp_data), self.get_child_count()), self.get_child_count()):
-            if i < self.get_child_count():
-                # Check missing required elements
-                valid &= self.get_child_node_by_idx(i).is_valid(None, errh)
-        return valid
-
-    def is_composite(self) -> bool:
-        """
-        :rtype: boolean
-        """
-        return True
+from ._element import element_if  # noqa: E402, I001
+from ._composite import composite_if  # noqa: E402, I001
 
 
 def load_map_file(map_file: str, param: Any, map_path: str | None = None) -> map_if:
